pyproject.toml
[tool.poetry]
name = "CryptoExchangEemulator"
version = "0.1.0"
description = ""
authors = ["Vladislav Marahovskiy <vladislav.marahovskiy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
litestar = {extras = ["standard"], version = "^2.16.0"}
python-dotenv = "^1.1.0"
loguru = "^0.7.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
